+++ src/GraphQLQueryManager.ts
@@ -1,13 +1,11 @@
 import {
-  QueueItem,
-  WhereArgument,
-  WhereExpression,
   FieldNodeInfo,
   LoaderOptions,
-  LoaderWhereExpression,
   QueryMeta,
   QueryPagination,
-  SearchOptions
+  QueueItem,
+  SearchOptions,
+  WhereExpression
 } from "./types";
 import { LoaderSearchMethod } from "./enums/LoaderSearchMethod";
 import { GraphQLInfoParser } from "./lib/GraphQLInfoParser";
@@ -195,7 +193,8 @@ export class GraphQLQueryManager {
         item.fields,
         entityManager.connection,
         queryBuilder,
-        alias
+        alias,
+        item.context
       );
       queryBuilder = this._addSelectFields(
         queryBuilder,